move _grab_total_items to v20, update README.rst
